[#5372] Convert Character sheet to ApplicationV2
Merges `ActorSheet5eCharacter` and `ActorSheet5eCharacter2` into
`CharacterActorSheet` and removes the old sheets, leaving behind
a small deprecation that extends the new sheets to avoid immediate
breakage.

Closes #5372
