+++ less/v2/actors.less
@@ -4,16 +4,6 @@
   /*  Outer Window                      */
   /* ---------------------------------- */
 
-  &::after {
-    content: "";
-    position: absolute;
-    pointer-events: none;
-    inset: 0;
-    border: var(--dnd5e-border-gold);
-    box-shadow: 0 0 20px var(--color-shadow-dark);
-    border-radius: 5px;
-  }
-
   .window-resizable-handle {
     bottom: 1px;
     right: 1px;
