+++ less/v2/apps.less
@@ -372,7 +372,7 @@
     }
   }
 
-  .gold-button {
+  .gold-button, &.standard-form button.gold-button {
     position: relative;
     border: none;
     border-radius: 4px;
@@ -452,13 +452,23 @@
     }
   }
 
-  a.deletion-control {
+  a.deletion-control, button.deletion-control {
     width: 32px;
     height: 32px;
     color: var(--dnd5e-color-red);
     display: flex;
-    justify-content: end;
+    align-items: flex-start;
+    justify-content: flex-end;
     font-size: var(--font-size-11);
+  }
+  button.deletion-control {
+    padding: 0;
+
+    &:focus-visible, &:hover:not(:disabled) {
+      background: none;
+      box-shadow: none;
+      color: var(--dnd5e-color-red);
+    }
   }
 
   hr {
