+++ less/v2/character.less
@@ -6,14 +6,24 @@
   --dnd5e-sheet-header-height: 170px;
   --dnd5e-sheet-header-right-width: 280px;
   --dnd5e-sheet-sidebar-width: 230px;
-  background: transparent;
-  min-width: 800px;
-  min-height: 700px;
+  background: var(--dnd5e-color-parchment);
   box-shadow: none;
 
   &.minimized {
     background: url("../../ui/denim075.png");
     box-shadow: 0 0 20px var(--color-shadow-dark);
+  }
+  &:not(.minimized, .minimizing, .maximizing) {
+    min-width: 800px;
+    min-height: 700px;
+  }
+
+  .window-content {
+    // TODO: Figure out a better way to do this
+    margin-block-start: -73px;
+    padding-block-end: 0;
+    gap: 0;
+    overflow: visible;
   }
 
   /* ---------------------------------- */
@@ -200,9 +210,10 @@
   /*  Sheet Body                        */
   /* ---------------------------------- */
 
-  .tab-details .tab-body { padding-top: 56px; }
+  &.tab-details .tab-body { padding-top: 56px; }
 
   .sheet-body {
+    flex: 1;
     border-radius: 5px;
     position: relative;
     background: url("ui/texture1.webp") no-repeat top center / auto 770px,
@@ -437,7 +448,10 @@
             align-items: center;
             justify-content: space-between;
 
-            > .config-button { color: var(--dnd5e-color-gold); }
+            > .config-button {
+              color: inherit;
+              font-size: inherit;
+            }
           }
         }
 
@@ -564,6 +578,7 @@
           }
 
           figure {
+            width: unset;
             position: relative;
             margin: 0;
           }
@@ -672,9 +687,9 @@
         }
       }
 
-      .tab:not(.details) { padding-inline: .5rem; }
+      .tab:not([data-tab="details"]) { padding-inline: .5rem; }
 
-      .tab:not(.details, .biography, .bastion) {
+      .tab:not([data-tab="details"], [data-tab="biography"], [data-tab="bastion"], [data-tab="specialTraits"]) {
         // enough room that the "create new" button doesn't overlap when scrolled all the way down
         padding-bottom: calc(1.5rem + 30px);
       }
@@ -794,7 +809,7 @@
     }
 
     /* Background Pills */
-    .details .pills-lg {
+    [data-tab="details"] .pills-lg {
       display: flex;
       flex-direction: column;
       gap: .5rem;
@@ -878,7 +893,7 @@
   }
 
   /* Editors */
-  .tab.biography > .bottom, .tab.bastion .description {
+  .tab[data-tab="biography"] > .bottom, .tab[data-tab="bastion"] .description {
     flex: 1;
     display: flex;
     flex-direction: column;
@@ -988,6 +1003,12 @@
         color: var(--dnd5e-color-black);
         font-size: var(--font-size-16);
         padding: 3px 0;
+
+        > button {
+          color: inherit;
+          font-size: inherit;
+          font-weight: inherit;
+        }
       }
 
       .score {
@@ -1051,19 +1072,19 @@
     }
   }
 
-  form:not(.tab-details) .ability-scores { display: none; }
+  &:not(.tab-details) .ability-scores { display: none; }
 
   /* ---------------------------------- */
   /*  Spells                            */
   /* ---------------------------------- */
 
-  .tab.spells.active {
+  .tab[data-tab="spells"].active {
     display: flex;
     flex-direction: column;
     gap: 1rem;
   }
 
-  .tab.spells .top {
+  .tab[data-tab="spells"] .top {
     display: flex;
     gap: 1rem;
     flex-wrap: wrap;
@@ -1085,7 +1106,7 @@
   /*  Biography                         */
   /* ---------------------------------- */
 
-  .tab.biography {
+  .tab[data-tab="biography"] {
     &.active {
       display: flex;
       flex-direction: column;
@@ -1137,6 +1158,7 @@
           height: unset;
           padding: .25rem;
           background: transparent;
+          border: 1px solid var(--color-border-light-tertiary);
         }
 
         .value {
@@ -1164,7 +1186,7 @@
   /*  Bastion                           */
   /* ---------------------------------- */
 
-  .tab.bastion {
+  .tab[data-tab="bastion"] {
     flex-direction: column;
     gap: 16px;
 
@@ -1357,7 +1379,7 @@
   /*  Collapsed                         */
   /* ---------------------------------- */
 
-  form.collapsed {
+  &.sidebar-collapsed {
     .main-content {
       grid-template-columns: 0 1fr;
       padding-left: 0;
@@ -1407,20 +1429,20 @@
       }
     }
 
-    .tab.biography > .bottom {
+    .tab[data-tab="biography"] > .bottom {
       padding: .5rem;
       .editor { min-height: unset; }
     }
 
     .main-content { display: unset; }
-    .sheet-body .tab-body .tab:not(.details) { padding: 0; }
+    .sheet-body .tab-body .tab:not([data-tab="details"]) { padding: 0; }
   }
 
   /* ---------------------------------- */
   /*  Edit Mode                         */
   /* ---------------------------------- */
 
-  form.editable {
+  &.editable {
     .ability-scores .ability-score .score { padding: 0; }
 
     :is(.skills, .tools, .saves) {
