+++ module/applications/actor/_module.mjs
@@ -1,5 +1,4 @@
-export {default as ActorSheet5eCharacter} from "./character-sheet.mjs";
-export {default as ActorSheet5eCharacter2} from "./character-sheet-2.mjs";
+export {default as CharacterActorSheet} from "./character-sheet.mjs";
 export {default as ActorSheet5eVehicle} from "./vehicle-sheet.mjs";
 export {default as GroupActorSheet} from "./group-sheet.mjs";
 export {default as NPCActorSheet} from "./npc-sheet.mjs";
@@ -34,6 +33,7 @@ export {default as ShortRestDialog} from "./rest/short-rest-dialog.mjs";
 // Deprecated
 export {default as BaseConfigSheet} from "./deprecated/base-config.mjs";
 export {default as ActorSheet5e} from "./deprecated/base-sheet.mjs";
+export * from "./deprecated/character-sheet.mjs";
 export * from "./deprecated/npc-sheet.mjs";
 export {default as ActorSheetMixin} from "./deprecated/sheet-mixin.mjs";
 export {default as ActorSheetV2Mixin} from "./deprecated/sheet-v2-mixin.mjs";
