+++ templates/actors/tabs/creature-spells.hbs
@@ -14,7 +14,7 @@
                 <span class="spellcasting-ability roboto-upper">{{ ability.label }}</span>
                 {{/if}}
                 {{else}}
-                <button type="button" class="radio-button" data-action="spellcasting"
+                <button type="button" class="radio-button" data-action="setSpellcastingAbility"
                         data-tooltip="DND5E.SpellAbilitySet" aria-pressed="{{#if primary}}true{{else}}false{{/if}}"
                         aria-label="{{ localize "DND5E.SpellAbilitySet" }}"></button>
                 {{/if}}
@@ -103,7 +103,9 @@
 
                 {{!-- Section Header --}}
                 <div class="items-header header">
-                    <h3 class="item-name spell-header">{{ localize label }}</h3>
+                    <h3 class="item-name spell-header {{~#if @root.isCharacter}} draggable{{/if}}">
+                        {{ localize label }}
+                    </h3>
                     <div class="item-header item-school">{{ localize "DND5E.SpellHeader.School" }}</div>
                     <div class="item-header item-usage">{{ localize "DND5E.SpellHeader.Time" }}</div>
                     <div class="item-header item-range">{{ localize "DND5E.SpellHeader.Range" }}</div>
